+++ HW 21/account.html
@@ -38,6 +38,7 @@
                     </tr>
                 </table>
             </div>
+            <p class="empty">Order history is empty</p>
             <div class="user-info">
                 <p class="title">My Info</p>
                 <div class="name">
@@ -50,6 +51,6 @@
             </div>
         </div>
     </main>
-    <script src="script.js"></script>
+    <script src="index.js" type="module"></script>
 </body>
 </html>
