+++ HW 21/index.html
@@ -32,7 +32,7 @@
         </div>
 
     </main>
-    <script src="script.js"></script>
+    <script src="index.js" type="module"></script>
 </body>
 
 </html>
