+++ HW 21/login.html
@@ -47,12 +47,13 @@
                     <input type="text" name="full-name" placeholder="Full name" required><br>
                     <input type="email" name="email" placeholder="Email Address" required><br>
                     <input type="password" name="password" placeholder="Password" autocomplete="on" required><br>
-                    <input type="password" name="verify-password" placeholder="Verify password" autocomplete="on" required><br>
+                    <input type="password" name="verify-password" placeholder="Verify password" autocomplete="on"
+                        required><br>
                     <button class="btn btn__register" type="submit">Create account</button>
                 </form>
             </div>
         </main>
-        <script src="script.js"></script>
+        <script src="index.js" type="module"></script>
     </body>
 </body>
