+++ HW 21/shoppingCart.html
@@ -27,7 +27,7 @@
         </nav>
     </header>
     <main class="main">
-        <p class="empty-cart">Your shopping cart is empty</p>
+        <p class="empty">Your shopping cart is empty</p>
         <div class="cart_wrapper">
             <div class="items-list">
                 <p class="title">Items in Shopping Cart</p>
@@ -52,7 +52,7 @@
             </div>
         </div>
     </main>
-    <script src="script.js"></script>
+    <script src="index.js" type="module"></script>
 </body>
 
 </html>
